QUAN LY TU VI/TAI_LIEU_DU_AN/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
openai>=1.40.0,<2
psycopg[binary]>=3.2,<4
psycopg-pool>=3.2,<4
lunar-vn>=1.1.1
pytest>=8.0,<9
python-dotenv>=1.0,<2
httpx>=0.27,<0.29
streamlit>=1.40,<2
jinja2>=3.1,<4
